Move compaing classes and migrate vector code to srgbcompaning
