+++ src/ImageSharp/ColorSpaces/Companding/GammaCompanding.cs
@@ -4,32 +4,33 @@
 using System;
 using System.Runtime.CompilerServices;
 
-namespace SixLabors.ImageSharp.ColorSpaces.Conversion.Implementation
+namespace SixLabors.ImageSharp.ColorSpaces.Companding
 {
     /// <summary>
-    /// Implements sRGB companding
+    /// Implements gamma companding
     /// </summary>
     /// <remarks>
-    /// For more info see:
     /// <see href="http://www.brucelindbloom.com/index.html?Eqn_RGB_to_XYZ.html"/>
     /// <see href="http://www.brucelindbloom.com/index.html?Eqn_XYZ_to_RGB.html"/>
     /// </remarks>
-    public static class SRgbCompanding
+    public static class GammaCompanding
     {
         /// <summary>
         /// Expands a companded channel to its linear equivalent with respect to the energy.
         /// </summary>
-        /// <param name="channel">The channel value</param>
+        /// <param name="channel">The channel value.</param>
+        /// <param name="gamma">The gamma value.</param>
         /// <returns>The <see cref="float"/> representing the linear channel value.</returns>
         [MethodImpl(InliningOptions.ShortMethod)]
-        public static float Expand(float channel) => channel <= 0.04045F ? channel / 12.92F : MathF.Pow((channel + 0.055F) / 1.055F, 2.4F);
+        public static float Expand(float channel, float gamma) => MathF.Pow(channel, gamma);
 
         /// <summary>
         /// Compresses an uncompanded channel (linear) to its nonlinear equivalent.
         /// </summary>
-        /// <param name="channel">The channel value</param>
+        /// <param name="channel">The channel value.</param>
+        /// <param name="gamma">The gamma value.</param>
         /// <returns>The <see cref="float"/> representing the nonlinear channel value.</returns>
         [MethodImpl(InliningOptions.ShortMethod)]
-        public static float Compress(float channel) => channel <= 0.0031308F ? 12.92F * channel : (1.055F * MathF.Pow(channel, 0.416666666666667F)) - 0.055F;
+        public static float Compress(float channel, float gamma) => MathF.Pow(channel, 1 / gamma);
     }
 }
