+++ src/ImageSharp/ColorSpaces/Companding/LCompanding.cs
@@ -3,17 +3,19 @@
 
 using System;
 using System.Runtime.CompilerServices;
+using SixLabors.ImageSharp.ColorSpaces.Conversion;
 
-namespace SixLabors.ImageSharp.ColorSpaces.Conversion.Implementation
+namespace SixLabors.ImageSharp.ColorSpaces.Companding
 {
     /// <summary>
-    /// Implements Rec. 2020 companding function (for 12-bits).
+    /// Implements L* companding
     /// </summary>
     /// <remarks>
-    /// <see href="http://en.wikipedia.org/wiki/Rec._2020"/>
-    /// For 10-bits, companding is identical to <see cref="Rec709Companding"/>
+    /// For more info see:
+    /// <see href="http://www.brucelindbloom.com/index.html?Eqn_RGB_to_XYZ.html"/>
+    /// <see href="http://www.brucelindbloom.com/index.html?Eqn_XYZ_to_RGB.html"/>
     /// </remarks>
-    public static class Rec2020Companding
+    public static class LCompanding
     {
         /// <summary>
         /// Expands a companded channel to its linear equivalent with respect to the energy.
@@ -22,15 +24,15 @@ namespace SixLabors.ImageSharp.ColorSpaces.Conversion.Implementation
         /// <returns>The <see cref="float"/> representing the linear channel value.</returns>
         [MethodImpl(InliningOptions.ShortMethod)]
         public static float Expand(float channel)
-            => channel < 0.08145F ? channel / 4.5F : MathF.Pow((channel + 0.0993F) / 1.0993F, 2.222222F);
+            => channel <= 0.08 ? 100 * channel / CieConstants.Kappa : ImageMaths.Pow3((channel + 0.16F) / 1.16F);
 
         /// <summary>
         /// Compresses an uncompanded channel (linear) to its nonlinear equivalent.
         /// </summary>
-        /// <param name="channel">The channel value.</param>
+        /// <param name="channel">The channel value</param>
         /// <returns>The <see cref="float"/> representing the nonlinear channel value.</returns>
         [MethodImpl(InliningOptions.ShortMethod)]
         public static float Compress(float channel)
-            => channel < 0.0181F ? 4500F * channel : (1.0993F * channel) - 0.0993F;
+            => channel <= CieConstants.Epsilon ? channel * CieConstants.Kappa / 100F : MathF.Pow(1.16F * channel, 0.3333333F) - 0.16F;
     }
 }
