+++ src/ImageSharp/ColorSpaces/Companding/Rec2020Companding.cs
@@ -4,33 +4,33 @@
 using System;
 using System.Runtime.CompilerServices;
 
-namespace SixLabors.ImageSharp.ColorSpaces.Conversion.Implementation
+namespace SixLabors.ImageSharp.ColorSpaces.Companding
 {
     /// <summary>
-    /// Implements gamma companding
+    /// Implements Rec. 2020 companding function (for 12-bits).
     /// </summary>
     /// <remarks>
-    /// <see href="http://www.brucelindbloom.com/index.html?Eqn_RGB_to_XYZ.html"/>
-    /// <see href="http://www.brucelindbloom.com/index.html?Eqn_XYZ_to_RGB.html"/>
+    /// <see href="http://en.wikipedia.org/wiki/Rec._2020"/>
+    /// For 10-bits, companding is identical to <see cref="Rec709Companding"/>
     /// </remarks>
-    public static class GammaCompanding
+    public static class Rec2020Companding
     {
         /// <summary>
         /// Expands a companded channel to its linear equivalent with respect to the energy.
         /// </summary>
         /// <param name="channel">The channel value.</param>
-        /// <param name="gamma">The gamma value.</param>
         /// <returns>The <see cref="float"/> representing the linear channel value.</returns>
         [MethodImpl(InliningOptions.ShortMethod)]
-        public static float Expand(float channel, float gamma) => MathF.Pow(channel, gamma);
+        public static float Expand(float channel)
+            => channel < 0.08145F ? channel / 4.5F : MathF.Pow((channel + 0.0993F) / 1.0993F, 2.222222F);
 
         /// <summary>
         /// Compresses an uncompanded channel (linear) to its nonlinear equivalent.
         /// </summary>
         /// <param name="channel">The channel value.</param>
-        /// <param name="gamma">The gamma value.</param>
         /// <returns>The <see cref="float"/> representing the nonlinear channel value.</returns>
         [MethodImpl(InliningOptions.ShortMethod)]
-        public static float Compress(float channel, float gamma) => MathF.Pow(channel, 1 / gamma);
+        public static float Compress(float channel)
+            => channel < 0.0181F ? 4500F * channel : (1.0993F * channel) - 0.0993F;
     }
 }
