+++ src/ImageSharp/ColorSpaces/RgbWorkingSpaces.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
+using SixLabors.ImageSharp.ColorSpaces.Companding;
 using SixLabors.ImageSharp.ColorSpaces.Conversion.Implementation;
 
 // ReSharper disable InconsistentNaming
