+++ src/ImageSharp/Processing/Processors/Convolution/Convolution2DProcessor.cs
@@ -102,7 +102,7 @@ namespace SixLabors.ImageSharp.Processing.Processors.Convolution
                                             int offsetX = x + fxr;
 
                                             offsetX = offsetX.Clamp(0, maxX);
-                                            Vector4 currentColor = sourceOffsetRow[offsetX].ToVector4().Premultiply();
+                                            Vector4 currentColor = Vector4Utils.Premultiply(sourceOffsetRow[offsetX].ToVector4());
 
                                             if (fy < kernelXHeight)
                                             {
@@ -128,7 +128,7 @@ namespace SixLabors.ImageSharp.Processing.Processors.Convolution
 
                                     ref TPixel pixel = ref targetRow[x];
                                     pixel.PackFromVector4(
-                                        new Vector4(red, green, blue, sourceRow[x].ToVector4().W).UnPremultiply());
+                                      Vector4Utils.UnPremultiply(new Vector4(red, green, blue, sourceRow[x].ToVector4().W)));
                                 }
                             }
                         });
